Implement new infrastructure (2)

+++ src/Model/BaseModel.php
@@ -7,6 +7,7 @@ use Isswp101\Persimmon\Contracts\Jsonable;
 use Isswp101\Persimmon\Contracts\Stringable;
 use Isswp101\Persimmon\DI\Container;
 use Isswp101\Persimmon\Traits\Containerable;
+use Isswp101\Persimmon\Traits\Eventable;
 use Isswp101\Persimmon\Traits\Timestampable;
 use JsonSerializable;
 
@@ -16,13 +17,13 @@ use JsonSerializable;
  * @property string updated_at
  *
  * @TODO
- * 1. Events
+ * 1. Events +
  * 2. Timestamps +
  * 3. Cache
  */
 abstract class BaseModel implements ModelInterface, Arrayable, Jsonable, Stringable, JsonSerializable
 {
-    use Containerable, Timestampable;
+    use Containerable, Timestampable, Eventable;
 
     protected $exists = false;
     protected $timestamps = false;
@@ -55,6 +56,9 @@ abstract class BaseModel implements ModelInterface, Arrayable, Jsonable, Stringa
 
     public function save(array $columns = null)
     {
+        if ($this->saving() === false) {
+            return;
+        }
         $repository = $this->di()->getRepository();
         if ($this->timestamps) {
             $this->updateTimestamps();
@@ -65,11 +69,20 @@ abstract class BaseModel implements ModelInterface, Arrayable, Jsonable, Stringa
             $repository->update($this);
         }
         $this->exists = true;
+        if ($this->saved() === false) {
+            return;
+        }
     }
 
     public function delete()
     {
+        if ($this->deleting() === false) {
+            return;
+        }
         $this->di()->getRepository()->delete($this->getId());
         $this->exists = false;
+        if ($this->deleted() === false) {
+            return;
+        }
     }
 }
